+++ Runtime/DataTypes/DatabaseTokenData.cs
@@ -1,0 +1,14 @@
+namespace EMullen.PlayerMgmt 
+{
+
+    [RequireVisibilityHandler(Handler.OWNER_AND_SERVER)]
+    public class DatabaseTokenData : PlayerDataClass 
+    {
+        public readonly string token;
+
+        public DatabaseTokenData(string token) 
+        {
+            this.token = token;
+        }
+    }
+}